feat(home): refactor code , add files to widget directory and add localization .
