+++ lib/shared_layers/localization/l10n_json_files/ar.json
@@ -52,5 +52,10 @@
   "home": "الرئيسية",
   "categories": "الفئات",
   "cart": "عربة التسوق",
-  "profile": "الملف الشخصي"
+  "profile": "الملف الشخصي",
+  "search": "بحث",
+  "viewAll": "عرض الكل",
+  "bestSeller": "الأكثر مبيعاً",
+  "occasions": "مناسبات"
+
 }
