+++ lib/shared_layers/localization/l10n_json_files/en.json
@@ -54,5 +54,9 @@
   "forgetPassword": "Forget Password",
   "dontHaveAccount": "Don't have an account?",
   "signUp": "Sign up",
-  "loginSuccess": "LoginSuccess"
+  "loginSuccess": "LoginSuccess",
+  "search": "Search",
+  "viewAll": "View All",
+  "bestSeller": "Best Seller",
+  "occasions": "Occasions"
 }
